data preprocess performance improved, with a slight precision problem.

+++ datasets/Chengdu/data_preprocess.py
@@ -6,18 +6,19 @@ import time
 from typing import Optional
 from tqdm import tqdm
 
+final_csv_header = ['vehicle_id', 'time', 'longitude', 'latitude', 'timestamp', 'x', 'y']
 
-def construct_dataframe_for_time_larger_than_0(df, new_row):
+
+def construct_dataframe_for_time_larger_than_0(new_row):
     list_of_rows = []
     for time_index in range(int(new_row.time)):
-        list_of_rows.append({
-            'vehicle_id': new_row.vehicle_id,
-            'time': time_index,
-            'longitude': new_row.longitude,
-            'latitude': new_row.latitude}
-        )
-    df = pd.concat([df, pd.DataFrame(list_of_rows)], axis=0, ignore_index=True)
-    return df
+        new_dict = {}
+        for item in ['vehicle_id', 'longitude', 'latitude', 'x', 'y']:
+            new_dict[item] = getattr(new_row, item)
+        new_dict['time'] = time_index
+        new_dict['timestamp'] = new_row.timestamp - int(new_row.time) + time_index
+        list_of_rows.append(new_dict)
+    return list_of_rows
 
 
 def trans_form_of_lat(lng: float, lat: float):
@@ -57,8 +58,8 @@ def get_distance(lng1: float, lat1: float, lng2: float, lat2: float) -> float:
 
 def generate_rows(old_row, new_row, add_number, no_long_lat=False):
     if no_long_lat:
-        add_longitude = 1
-        add_latitude = 1
+        add_longitude = 0
+        add_latitude = 0
     else:
         add_longitude = (new_row.x - old_row.x) / add_number
         add_latitude = (new_row.y - old_row.y) / add_number
@@ -67,6 +68,7 @@ def generate_rows(old_row, new_row, add_number, no_long_lat=False):
         list_of_rows.append({
             'vehicle_id': old_row.vehicle_id,
             'time': int(old_row.time) + time_index + 1,
+            'timestamp': old_row.timestamp + time_index + 1,
             'longitude': old_row.longitude,
             'latitude': old_row.latitude,
             'x': float(old_row.x) + (time_index + 1) * add_longitude,
@@ -181,21 +183,21 @@ class VehicleTrajectoriesProcessor(object):
             out_file=self._out_file
         )
 
-    def process_row(self, df, old_row, new_row):
+    def process_row(self, old_row, new_row):
         if old_row.vehicle_id == new_row.vehicle_id:
             add_number = int(new_row.time) - int(old_row.time) - 1
             if add_number > 0:
                 list_of_rows = generate_rows(old_row, new_row, add_number)
-                df = pd.concat([df, pd.DataFrame(list_of_rows)], axis=0, ignore_index=True)
+            else:
+                list_of_rows = []
         else:
             if int(old_row.time) < self._time_end_int - self._time_start_int:
                 list_of_rows = generate_rows(old_row, old_row,
                                              int(self._time_end_int - self._time_start_int) - int(old_row.time) - 1,
                                              no_long_lat=True)
-                df = pd.concat([df, pd.DataFrame(list_of_rows)], axis=0, ignore_index=True)
             if int(new_row.time) > 0:
-                df = construct_dataframe_for_time_larger_than_0(df, new_row)
-        return df
+                list_of_rows.extend(construct_dataframe_for_time_larger_than_0(new_row))
+        return list_of_rows
 
     def process(
             self,
@@ -206,15 +208,16 @@ class VehicleTrajectoriesProcessor(object):
     ) -> None:
 
         # if files are already cached, skip reading
-        if not (os.path.exists(out_file + "_without_fill" + ".csv") and os.path.exists(out_file + ".csv")):
-            print("Reading data...")
-            df = pd.read_csv(
+        if not (os.path.exists(out_file + "_without_fill" + ".csv")):
+
+            reader = tqdm(pd.read_csv(
                 self._file_name,
                 names=['vehicle_id', 'order_number', 'time', 'longitude', 'latitude'],
-                dtype={'vehicle_id': str, 'order_number': str, 'time': np.int32,
-                       'longitude': np.float64, 'latitude': np.float64},
-                header=0
-            )
+                # dtype={'vehicle_id': str, 'order_number': str, 'time': 'UInt64',
+                #        'longitude': 'Float64', 'latitude': 'Float64'},
+                header=0, chunksize=1000
+            ), desc='Loading CSV data')
+            df = pd.concat([chunk for chunk in reader])
             # 经纬度定位
             df.drop(df.columns[[1]], axis=1, inplace=True)
             df.dropna(axis=0)
@@ -235,6 +238,7 @@ class VehicleTrajectoriesProcessor(object):
             # sorted
             df.sort_values(by=['vehicle_id', 'time'], inplace=True, ignore_index=True)
 
+            # check if file without_fill exist
             print("Discretize data...")
             vehicle_number = 0
             old_vehicle_id = None
@@ -258,18 +262,26 @@ class VehicleTrajectoriesProcessor(object):
 
             if self._output_analysis:
                 df.to_csv(out_file + "_without_fill" + ".csv")
-
+        else:
+            df = pd.read_csv(out_file + "_without_fill" + ".csv", header=0)
+        if not os.path.exists(out_file + ".csv"):
             print("Filling data...")
             old_row = None
+            all_list_of_rows = []
             for row in tqdm(df.itertuples(index=False), total=len(df)):
                 new_row = row
                 if old_row:
-                    df = self.process_row(df, old_row, new_row)
+                    list_of_rows = self.process_row(old_row, new_row)
                     old_row = new_row
                 else:
                     if int(new_row.time) > 0:
-                        df = construct_dataframe_for_time_larger_than_0(df, new_row)
+                        list_of_rows = construct_dataframe_for_time_larger_than_0(new_row)
+                    else:
+                        list_of_rows = []
                     old_row = new_row
+                all_list_of_rows.extend(list_of_rows)
+            # jesus, you should cat df, not that original_df...
+            df = pd.concat([df, pd.DataFrame(all_list_of_rows)], axis=0, ignore_index=True)
             df.sort_values(by=['vehicle_id', 'time'], inplace=True, ignore_index=True)
             df.to_csv(out_file + ".csv")
 
@@ -277,7 +289,7 @@ class VehicleTrajectoriesProcessor(object):
             analyst = VehicleTrajectoriesAnalyst(
                 trajectories_file_name=out_file + ".csv",
                 trajectories_file_name_with_no_fill=out_file + "_without_fill" + ".csv",
-                during_time=self._time_start_int - self._time_start_int,
+                during_time=self._time_end_int - self._time_start_int,
             )
             analyst.output_characteristics()
 
@@ -285,7 +297,7 @@ class VehicleTrajectoriesProcessor(object):
         row['vehicle_id'] = vehicle_number
         longitude, latitude = gcj02_to_wgs84(float(row['longitude']), float(row['latitude']))
         row['timestamp'] = row['time']
-        row['time'] -= self._time_start_int
+        row['time'] = int(row['time'] - self._time_start_int)
         x = get_distance(longitude_min, latitude_min, longitude, latitude_min)
         y = get_distance(longitude_min, latitude_min, longitude_min, latitude)
         for item in ['x', 'y', 'longitude', 'latitude']:
@@ -333,8 +345,7 @@ class VehicleTrajectoriesAnalyst(object):
         self._during_time = during_time
 
     def output_characteristics(self):
-
-        df = pd.read_csv(self._trajectories_file_name, names=['vehicle_id', 'time', 'x', 'y'], header=0)
+        df = pd.read_csv(self._trajectories_file_name, names=final_csv_header, header=0)
         vehicle_ids = df['vehicle_id'].unique()
         number_of_vehicles_in_seconds = np.zeros(self._during_time, dtype=np.int32)
         vehicle_dwell_times = []
@@ -363,7 +374,7 @@ class VehicleTrajectoriesAnalyst(object):
 
         vehicle_speeds = []
         df = pd.read_csv(self._trajectories_file_name_with_no_fill,
-                         names=['vehicle_id', 'time', 'x', 'y'], header=0)
+                         names=final_csv_header, header=0)
         vehicle_ids = df['vehicle_id'].unique()
 
         for vehicle_id in vehicle_ids:
@@ -382,7 +393,7 @@ class VehicleTrajectoriesAnalyst(object):
                 speed = distance / (row.time - last_time)
                 if not np.isnan(speed):
                     vehicle_speed.append(speed)
-                last_time = time
+                last_time = row.time
                 last_x = row.x
                 last_y = row.y
             if vehicle_speed:
@@ -395,34 +406,35 @@ class VehicleTrajectoriesAnalyst(object):
         print("Standard deviation of speed (m/s):", asv_std)
 
 
+parent_dir_name = os.path.join("/workspace", "saved_data", 'datasets', 'Chengdu_taxi')
+trajectory_file_name: str = os.path.join(parent_dir_name, 'gps_20161116')
+longitude_min: float = 104.04565967220308
+latitude_min: float = 30.654605745741608
+start_hour = 22
+start_minute = 0
+start_second = 0
+end_hour = 22
+end_minute = 5
+end_second = 0
+padded_start_hour = str(start_hour).zfill(2)
+padded_start_minute = str(start_minute).zfill(2)
+trajectories_time_start: str = ('2016-11-16' + ' ' + padded_start_hour + ':' +
+                                padded_start_minute + ':' + str(start_second).zfill(2))
+padded_end_hour = str(end_hour).zfill(2)
+padded_end_minute = str(end_minute).zfill(2)
+trajectories_time_end: str = '2016-11-16' + ' ' + padded_end_hour + ':' + \
+                             padded_end_minute + ':' + str(end_second).zfill(2)
+
+trajectories_out_file_name: str = os.path.join(parent_dir_name,
+                                               f'trajectories_20161116_{padded_start_hour}'
+                                               f'{padded_start_minute}_{padded_end_hour}'
+                                               f'{padded_end_minute}')
+
 if __name__ == "__main__":
     """Vehicle Trajectories Processor related."""
-    trajectories_file_name: str = os.path.join("/workspace", "saved_data", 'datasets',
-                                               'Chengdu_taxi', 'gps_20161116')
-    longitude_min: float = 104.04565967220308
-    latitude_min: float = 30.654605745741608
-    start_hour = 22
-    start_minute = 0
-    start_second = 0
-    end_hour = 22
-    end_minute = 5
-    end_second = 0
-    padded_start_hour = str(start_hour).zfill(2)
-    padded_start_minute = str(start_minute).zfill(2)
-    trajectories_time_start: str = ('2016-11-16' + ' ' + padded_start_hour + ':' +
-                                    padded_start_minute + ':' + str(start_second).zfill(2))
-    padded_end_hour = str(end_hour).zfill(2)
-    padded_end_minute = str(end_minute).zfill(2)
-    trajectories_time_end: str = '2016-11-16' + ' ' + padded_end_hour + ':' + \
-                                 padded_end_minute + ':' + str(end_second).zfill(2)
-    trajectories_out_file_name: str = os.path.join("/workspace", "saved_data", 'datasets',
-                                                   'Chengdu_taxi',
-                                                   f'trajectories_20161116_{padded_start_hour}'
-                                                   f'{padded_start_minute}_{padded_end_hour}'
-                                                   f'{padded_end_minute}')
-
+ 
     processor = VehicleTrajectoriesProcessor(
-        file_name=trajectories_file_name,
+        file_name=trajectory_file_name,
         longitude_min=longitude_min,
         latitude_min=latitude_min,
         map_width=2000.0,
